requirements.txt
# STAC & Geospatial stuff
planetary_computer
pystac_client
stackstac
rasterio
rioxarray
geopandas
shapely
leafmap

# Jupyter
ipykernel

# General
numpy
requests
pandas

# utilities
patool

# Dask
dask
bokeh==2.4.2

# Plotting & Images
Pillow
scikit-image
seaborn

# Optionally, install the segment-geospatial library (it is very large, ~7GB)
segment-geospatial
# groundingdino-py 
# localtileserver 
# jupyter-server-proxy
